use int

+++ MingdaoSDKClass/Model/MDCalendarCategory.h
@@ -12,7 +12,7 @@
 
 @property (strong, nonatomic) NSString *catID;
 @property (strong, nonatomic) NSString *catName;
-@property (assign, nonatomic) NSInteger displayOrder;
+@property (assign, nonatomic) int displayOrder;
 @property (assign, nonatomic) int color;
 
 - (MDCalendarCategory *)initWithDictionary:(NSDictionary *)aDic;
